// src/FieldForm2/interface.ts
export type InternalNamePath = (string | number)[];
export type NamePath = string | number | InternalNamePath;

export type StoreValue = any;
export interface Store {
  [name: string]: StoreValue;
}

export interface Meta {
  touched: boolean;
  validating: boolean;
  errors: string[];
  name: InternalNamePath;
}

export interface InternalFieldData extends Meta {
  value: StoreValue;
}

/**
 * Used by `setFields` config
 */
export interface FieldData extends Partial<Omit<InternalFieldData, 'name'>> {
  name: NamePath;
}

export type RuleType =
  | 'string'
  | 'number'
  | 'boolean'
  | 'method'
  | 'regexp'
  | 'integer'
  | 'float'
  | 'object'
  | 'enum'
  | 'date'
  | 'url'
  | 'hex'
  | 'email';

type Validator = (
  rule: RuleObject,
  value: StoreValue,
  callback: (error?: string) => void,
) => Promise<void> | void;

export type RuleRender = (form: FormInstance) => RuleObject;
// ...
export type RuleObject = AggregationRule | ArrayRule;

export type Rule = RuleObject | RuleRender;

export interface ValidateErrorEntity<Values = any> {
  values: Values;
  errorFields: { name: InternalNamePath; errors: string[] }[];
  outOfDate: boolean;
}

export interface FieldEntity {
  // onStoreChange: (
  //   store: Store,
  //   namePathList: InternalNamePath[] | null,
  //   info: ValuedNotifyInfo,
  // ) => void;
  onStoreChange: () => void;
  // isFieldTouched: () => boolean;
  // isFieldDirty: () => boolean;
  // isFieldValidating: () => boolean;
  validateRules: () => any;
  // getMeta: () => Meta;
  // getNamePath: () => InternalNamePath;
  // getErrors: () => string[];
  props: {
    name?: string;
    rules?: Rule[];
    dependencies?: NamePath[];
    initialValue?: any;
  };
}

export interface FieldError {
  name: InternalNamePath;
  errors: string[];
}
// ...
export interface Callbacks<Values = any> {
  onValuesChange?: (changedValues: any, values: Values) => void;
  onFieldsChange?: (changedFields: FieldData[], allFields: FieldData[]) => void;
  onFinish?: (values: Values) => void;
  onFinishFailed?: (errorInfo: ValidateErrorEntity<Values>) => void;
}

export interface InternalHooks {
  // dispatch: (action: ReducerAction) => void;
  registerField: (entity: FieldEntity) => () => void;
  // useSubscribe: (subscribable: boolean) => void;
  setInitialValues: (values: Store, init: boolean) => void;
  setCallbacks: (callbacks: Callbacks) => void;

  // getFields: (namePathList?: InternalNamePath[]) => FieldData[];
  // setValidateMessages: (validateMessages: ValidateMessages) => void;
  // setPreserve: (preserve?: boolean) => void;
}

export interface FormInstance<Values = any> {
  // Origin Form API
  getFieldValue: (name: string) => StoreValue;
  // getFieldValue: (name: NamePath) => StoreValue;
  getFieldsValue(): Values;
  // getFieldsValue(nameList: NamePath[] | true, filterFunc?: (meta: Meta) => boolean): any;
  // getFieldError: (name: NamePath) => string[];
  // getFieldsError: (nameList?: NamePath[]) => FieldError[];
  // isFieldsTouched(nameList?: NamePath[], allFieldsTouched?: boolean): boolean;
  // isFieldsTouched(allFieldsTouched?: boolean): boolean;
  // isFieldTouched: (name: NamePath) => boolean;
  // isFieldValidating: (name: NamePath) => boolean;
  // isFieldsValidating: (nameList: NamePath[]) => boolean;
  resetFields: (fields?: string[]) => void;
  // setFields: (fields: FieldData[]) => void;
  setFieldsValue: (value: any) => void;
  // setFieldsValue: (value: RecursivePartial<Values>) => void;
  validateFields: any;

  // New API
  submit: () => void;
}